moved mesh-related functions in mesh.go
